+++ src/main/java/org/bilanzius/persistence/sql/adapter/SqlTransactionAdapter.java
@@ -11,6 +11,6 @@ public class SqlTransactionAdapter implements SqlDataAdapter<Transaction> {
     public Transaction deserialize(ResultSet resultSet) throws SQLException {
         return new Transaction(resultSet.getInt("id"), resultSet.getInt("userId"),
                 resultSet.getInt("accountId"), resultSet.getInt("categoryId"),
-                resultSet.getDouble("money"), resultSet.getString("description"));
+                resultSet.getBigDecimal("money"), resultSet.getString("description"));
     }
 }
